+++ setup.cfg
@@ -5,6 +5,8 @@ extend-ignore =
     E203,
     E401,
     # https://github.com/psf/black/issues/1029
+    # https://github.com/psf/black/issues/1029
+    E501,
     W503,
 exclude =
     srv/migrations
